[REFACTOR]: Divide page into small components
